requirements.txt
streamlit
tensorflow
gdown
opencv-python
Pillow
numpy
